Testing fade

+++ fadeColorPicker.cpp
@@ -44,7 +44,7 @@ HSLAPixel fadeColorPicker::operator()(point p)
     // int currX = p.x + borderSize;
     // int currY = p.y + borderSize;
 
-    double distance = sqrt(pow((int)p.x-(int)p.c.x, 2) + pow((int)p.y-(int)p.c.y, 2));
+    int distance = sqrt(pow((int)p.x-(int)p.c.x, 2) + pow((int)p.y-(int)p.c.y, 2));
     double l_new = p.c.color.l * pow(fadeFactor, distance);
     return HSLAPixel(p.c.color.h, p.c.color.s, l_new);
